requirements.txt
# =============================
# YOLOv8 Landmark Inference UI
# =============================

# Core YOLO and PyTorch dependencies
ultralytics==8.3.227
ultralytics-thop==2.0.18
torch>=2.0.0
torchvision>=0.15.0

# Image processing and array manipulation
opencv-python-headless==4.10.0.84
Pillow>=9.4.0
numpy>=1.24.0

# Web application framework
streamlit>=1.38.0

# Visualization and progress tracking
matplotlib>=3.7.0
tqdm>=4.66.0

# Optional utility for zip file handling
zipfile36>=0.1.3

# Notes:
# - If running on GPU, ensure CUDA-compatible PyTorch version is installed.
#   Example for CUDA 12.1:
#   torch==2.3.0+cu121
#   torchvision==0.18.0+cu121
#   --extra-index-url https://download.pytorch.org/whl/cu121
